// src/lib/kimai/models/activity-collection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kimai - API Docs
 * JSON API for the Kimai time-tracking software: [API documentation](https://www.kimai.org/documentation/rest-api.html), [Swagger definition file](doc.json) 
 *
 * The version of the OpenAPI document: 0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ActivityMeta } from './activity-meta';
// May contain unused imports in some cases
// @ts-ignore
import { Team } from './team';

/**
 * 
 * @export
 * @interface ActivityCollection
 */
export interface ActivityCollection {
    /**
     * 
     * @type {string}
     * @memberof ActivityCollection
     */
    'parentTitle'?: string;
    /**
     * 
     * @type {number}
     * @memberof ActivityCollection
     */
    'project'?: number;
    /**
     * 
     * @type {number}
     * @memberof ActivityCollection
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityCollection
     */
    'name': string;
    /**
     * 
     * @type {boolean}
     * @memberof ActivityCollection
     */
    'visible': boolean;
    /**
     * All visible meta (custom) fields registered with this activity
     * @type {Array<ActivityMeta>}
     * @memberof ActivityCollection
     */
    'metaFields'?: Array<ActivityMeta>;
    /**
     * If no team is assigned, everyone can access the activity
     * @type {Array<Team>}
     * @memberof ActivityCollection
     */
    'teams'?: Array<Team>;
    /**
     * 
     * @type {string}
     * @memberof ActivityCollection
     */
    'color'?: string;
}