// src/lib/kimai/models/timesheet-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kimai - API Docs
 * JSON API for the Kimai time-tracking software: [API documentation](https://www.kimai.org/documentation/rest-api.html), [Swagger definition file](doc.json) 
 *
 * The version of the OpenAPI document: 0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { TimesheetMeta } from './timesheet-meta';

/**
 * 
 * @export
 * @interface TimesheetEntity
 */
export interface TimesheetEntity {
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'activity'?: number;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'project'?: number;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'user'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof TimesheetEntity
     */
    'tags'?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof TimesheetEntity
     */
    'begin': string;
    /**
     * 
     * @type {string}
     * @memberof TimesheetEntity
     */
    'end'?: string;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'duration'?: number;
    /**
     * 
     * @type {string}
     * @memberof TimesheetEntity
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'rate'?: number;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'internalRate'?: number;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'fixedRate'?: number;
    /**
     * 
     * @type {number}
     * @memberof TimesheetEntity
     */
    'hourlyRate'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof TimesheetEntity
     */
    'exported': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof TimesheetEntity
     */
    'billable': boolean;
    /**
     * All visible meta (custom) fields registered with this timesheet
     * @type {Array<TimesheetMeta>}
     * @memberof TimesheetEntity
     */
    'metaFields'?: Array<TimesheetMeta>;
}